media_clipping_agent/requirements.txt
# Media Clipping Agent - Python Dependencies
# Install with: pip install -r requirements.txt

# Core dependencies
playwright>=1.40.0
openpyxl>=3.1.0
Pillow>=10.0.0
pandas>=2.0.0

# Optional but recommended
python-dateutil>=2.8.0
requests>=2.31.0
